Add flux-operator flag to disable updates
It has became a problem since version 0.5.2[1], where there was a change
to when how flux will run `helm update`. This is causing a problem with
fluentd-cloudwatch dependency as it's an incubator project...

We don't care for the updates done automatically by flux. We'd like to
disable this.

This however highlights the problem, that we have a hardcoded set of
charts for flux, meaning it's difficult to upgrade and maintain. Need to
address in a near future.

[1]: https://github.com/weaveworks/flux/releases/tag/helm-0.5.2