Add splunk assumable read-only role
Cyber would like to have an access to our AWS accounts and snoop around
for some common security missconfiguration.

This would create a role and a policy to allow this to happen.

It originated from [cyber-security-splunk-onboarding] repo, but has been
slightly amended to our needs.

Things that are different:
- the format of the policies - Terraform
instead of JSON for easier management and insight
- the "aws_iam_role_policy.splunk_aws_ro_policy" `name` field has been
changed from `policy` to `SplunkROPolicy`

Sorry (not sorry) for both :D I couldn't live without these changes.

[cyber-security-splunk-onboarding]: https://github.com/alphagov/cyber-security-splunk-onboarding/blob/master/splunk_aws_role.tf